Fixed a bug affecting upgrades from the "anon" scene to the "user" scene, as when we upgrade the user's assets to allow them to submit an event via the apostrophe-moderator module. This bug was typically reported as a minify bug but actually comes into play on any such upgrade; it's just that minify reveals it immediately at startup.
We now resend the entire CSS world on such upgrades. This is necessary because the diff we really would have to compute otherwise is on the CSS output, not the set of LESS files to be included, and that is not practical.

This is inefficient, but it only affects scene upgrades, which are fairly rare.

TODO: it would be good to remove the old stylesheets when adding the new one, note the right way involves the "disabled" property:

http://stackoverflow.com/questions/3182840/removing-or-replacing-a-stylesheet-a-link-with-javascript-jquery

However we'd have to look at whether this displays a nasty flash of unstyled content before proceeding. And we'd have to know which link elements came from our asset mechanism and leave any custom link elements alone.

Fixes https://github.com/punkave/apostrophe-blocks/issues/2