You can now edit relationships, which are now driven by standard schemas. They do not currently support joins. That is, you can't add another join in the "relationship" property of a field.
By default you get apostrophe-docs:relationshipEditor.html, which is a simple schema modal, and apostrophe-docs-relationship-editor, which is the corresponding moog type that powers it browser-side. You can set a custom relationshipTemplate property when defining the join, which should be a cross-module-include path. You can also set a relationshipEditor property, which should be a moog type name for your alternative implementation of the editor.