Pass root option to clean-css when minifying CSS
LESS will preserve @import statements for regular CSS files when
compiling files, which means that when minifying, clean-css is the
module  responsible for the final concatenation of all CSS files.
Before this change, the @import paths would be absolute, and
clean-css wouldn't know where to start looking. See #809 for more
details.