2.23.1:
cleanCss needs to know that the output CSS files are going to live in apos-minified in order to correctly parse @import
statements that pull in plain .css files.

Also, the mechanism for prefixing URLs in CSS code was not applied at the correct stage of the bundling process (the
minify stage), which broke the ability to reference fonts, images, etc. via URLs beginning with /modules when using
an S3  asset bundle.