Strengthen broken infinite scroll in some browsers and site configurations
Problem #1:

Some browsers (among them Safari) fails to find the scroll position from html / body using jQuery. This is potentially related to the site having `overflow-x: hidden;` on the `html`tag.

This PR strengthens the browser lookup for the scroll position in all browsers (that I was able to test).

Problem #2:
Additionally jQuery fails to find the `offset.top` of the hidden "load more" button, due the button being `display: none;` in CSS/LESS. According to the jQuery documentation an element being styled with `display: none;` is excluded from the rendering tree and thus has a position that is undefined. This has been changed to `display:block;` and `visibility:hidden` This can affect the invisible space on existing sites below the AJAX content, but there's no other way and site devs can work around this by decreasing the height of the element and/or its parent. The button and it's parent(s) is removed after all content has been added to the page (nothing changed here).

Additional improvements:

* The on scroll event was always attached, even on pages which did not have any AJAX features. This PR does a check for the `data-apos-ajax-infinite-scroll` and only attaches the event watcher in case there is such an element in the DOM.

* The aforementioned scroll event watcher was firing on every scroll and this can lead to significant slugginess on scroll, especially on devices with less power, such as mobile. I have wrapped the event in a lodash throttle function which ensures a scroll event is only sent every 200ms. This should be plenty, as the code operates with a decent offset (500px default). In my testing I did not notice any delays in fetching data.