Fix arrow position in filter dropdowns
This PR fixes the position of the dropdown arrow in filters. Basically there was a hardcoded `top` in the `::after` which took precedence against the recent update to all select dropdowns. In a perfect world the `right` position should also be in the "core" dropdown styling but I don't want to break more than I fix at this time.