Fix image re-crop when image has `options.minSize` config
This PR should fix re-cropping when `options.minSize` is set for the image.

I was reading the code and got somewhat reluctant to attempt a fix due to inline comments, but tried to enable `autoCrop: true;` just to see what happened. Didn't work. I then remembered my annoying #1392 which I had also noticed myself at an earlier point. The missing padding can be annoying when you have large images.

So then **magic** (maybe) happened. I went on to add a 20px padding on the `apos-mode-body` for the image crop editor and just suddenly the image crop started loading at the correct position. Well, at least most of the times. I noticed one time where I may have gotten a wrong position, but I was also opening and cloing the console.

Overall I think this is an improvement to the current situation. Even if the crop is not positioned at the right position when starting a re-crop, it will still have the correct ratio if `options.aspectRatio` is in use. Before this PR it would just load a blank image on re-crop, which meant you would have to start the process over completely.

Padding is set to `@apos-padding-2` which I think is plenty of space to grab the from the corners.