Remove .apos-area-widgets, so that we get a consistent DOM structure
Previously, an additional wrapper element would be added to areas.
That element would wrap the widgets contained in that area, but doesn't
serve a very big function, and makes it a lot harder for developers to
style their sites, since this particular wrapper would only appear for
logged in users and not for logged out users.

The one function it had seems to have been to help with dragging and
dropping, but it's not necessary for that to work.

It's possible that it's too optimistic to remove this wrapper altogether,
but this is a first attempt at doing so in order to start a discussion
around whether it's reasonable at all.