Change apostrophe-rich-text-widgets to apply first CKEDITOR style when editor starts up
This seems to previously have been sort of the intention, but this change
makes it so that developers can control the default style of the CKEDITOR
instance properly. However, this implementation is somewhat naive in that
it always picks the first style if there are styles that have been defined.

CKEDITOR will always provide a <p> default (ie. if I declare an <h2> as the
default style, start the CKEDITOR instance and then immediately press Enter,
I'm back in a <p> tag instead of <h2>). But this change is breaking in the
sense that previously, developers could have provided styles only for the text
elements that weren't the regular <p> tags, but still have learned to expect
that to be the default when adding a new apostrophe-rich-text-widget.

On the other hand, this change will only apply to apostrophe-rich-text-widgets
when they are added to the page (or technically when they initialize and have
no content at all). That means the default behavior changes, but only for
new widgets that are added to the page.

Perhaps the sensible way forward is still to let developers explicitly mark
a style as the default, what do you think?