Improvements to apos.images.srcset function
Firstly, this change makes it so that the apos.attachments.isSized
return value is respected. If a particular attachment isn't sizes
(eg. if it's an SVG), we return an empty srcset attribute. This change
was prompted by #1595.

Secondly, we now never return image sizes that are larger than the
original. And we also make sure to include the original image size
in the srcset return value.